// bike-parking-app/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  public: {
    Tables: {
      Favorites: {
        Row: {
          created_at: string;
          id: number;
          location_address: string | null;
          location_id: string | null;
          user_id: string;
          username: string | null;
          x_coord: number | null;
          y_coord: number | null;
        };
        Insert: {
          created_at?: string;
          id?: number;
          location_address?: string | null;
          location_id?: string | null;
          user_id?: string;
          username?: string | null;
          x_coord?: number | null;
          y_coord?: number | null;
        };
        Update: {
          created_at?: string;
          id?: number;
          location_address?: string | null;
          location_id?: string | null;
          user_id?: string;
          username?: string | null;
          x_coord?: number | null;
          y_coord?: number | null;
        };
        Relationships: [
          {
            foreignKeyName: "public_Favorites_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          }
        ];
      };
      Pending: {
        Row: {
          created_at: string | null;
          description: string | null;
          email: string | null;
          id: number;
          image: Json | null;
          request_type: string | null;
          x_coord: number | null;
          y_coord: number | null;
        };
        Insert: {
          created_at?: string | null;
          description?: string | null;
          email?: string | null;
          id?: number;
          image?: Json | null;
          request_type?: string | null;
          x_coord?: number | null;
          y_coord?: number | null;
        };
        Update: {
          created_at?: string | null;
          description?: string | null;
          email?: string | null;
          id?: number;
          image?: Json | null;
          request_type?: string | null;
          x_coord?: number | null;
          y_coord?: number | null;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
      PublicSchema["Views"])
  ? (PublicSchema["Tables"] &
      PublicSchema["Views"])[PublicTableNameOrOptions] extends {
      Row: infer R;
    }
    ? R
    : never
  : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
  ? PublicSchema["Enums"][PublicEnumNameOrOptions]
  : never;